FriendsList ✅
Currently:
-> Behaviour is good for now.
-> The code has been cleaned up.
-> JsonResponses have been built.

 Next step:
-> Test
